get agreement by name

+++ src/main/java/com/example/demo/Product.java
@@ -2,7 +2,7 @@ package com.example.demo;
 
 public class Product {
   public String name;
-  public Float price;
+  public Double price;
   public Product parent;
   public Product[] products;
 }
